refactor: modularize mol3D class into a package with io and vectorized geometry
