+++ molSimplify/Classes/mol3D/__init__.py
@@ -1,0 +1,1 @@
+from .core import mol3D
